% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{AutoPlotting}
\alias{AutoPlotting}
\title{Automatic plotting of Dual STATIS results}
\usage{
AutoPlotting(
  REF,
  ADD = NULL,
  BPS = NULL,
  plotIS = TRUE,
  plotCO = TRUE,
  plotPC = TRUE,
  plotBags = TRUE
)
}
\arguments{
\item{REF}{A Dual STATIS results object from reference tables.}

\item{ADD}{A Dual STATIS results object from additional tables.}

\item{BPS}{A list of bagplot hulls and centers for Interstructure and Intrastructure.}

\item{plotIS}{A logical value indicating whether the Interstructure must be plotted or not.}

\item{plotCO}{A logical value indicating whether the Intrastructure must be plotted or not.}

\item{plotPC}{A logical value indicating whether the Parallel Coordinates must be plotted or not.
If number of observations in Reftable or Addtable is greater than 250, a sample is picked.}

\item{plotBags}{A logical value indicating whether the bagplots must be plotted or not.}
}
\description{
Default routine for automatic plotting of reference batches and
testing batches as part of the DS-PC approach
}
\details{
This function creates graphs for Interstructure and Intrastructure
analysis of reference and/or additional batches
}
\examples{
#DS-PC PlasticBags data
data( PlasticBags )
REFtab = PlasticBags$Ref$data[1:1500,]
ADDtab = PlasticBags$Add$data
REFfac = PlasticBags$Ref$factor[1:1500]
ADDfac = PlasticBags$Add$factor
DSPCres = AutoProcessing(REFtab, REFfac, ADDtab, ADDfac)
AutoPlotting( DSPCres$REF, DSPCres$ADD, DSPCres$BPS )
}
