% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{AutoProcessing}
\alias{AutoProcessing}
\title{Automatic processing of data using Dual STATIS}
\usage{
AutoProcessing(
  RefTable,
  RefFactor,
  AddTable = NULL,
  AddFactor = NULL,
  centering = TRUE,
  scaling = TRUE,
  normalizing = TRUE
)
}
\arguments{
\item{RefTable}{A matrix or data frame containing the K reference tables stacked
downwards one after another. Is preferred that its columns and rows have names.}

\item{RefFactor}{A factor with the same number of entries as rows are in \code{RefTable}.
It is used to define the different tables involved in the analysis.}

\item{AddTable}{A matrix or data frame containing the tables of testing batches
stacked downwards one after another. Is preferred that its columns and rows have names.}

\item{AddFactor}{A factor with the same number of entries as rows are in \code{AddTable}.
It is used to define the different tables involved in the analysis.}

\item{centering}{either a logical indicating whether center the data or not (using global mean),
a value to perform the centering or a vector indicating a value for every variable.}

\item{scaling}{either a logical indicating whether scale the data or not (using global standard
deviations), a value to perform the centering or a vector indicating a value for every variable.}

\item{normalizing}{When TRUE, data is normalized dividing by the corresponding number of rows
in each table. If FALSE, the tables are not normalized. Optionally, a function which takes
each table as argument can be given, the value obtained will be divided from each corresponding table.}

\item{FactOnTable}{if \code{TRUE}, \code{RefFactor} and/or \code{AddFactor} are defined
as the first column  of \code{RefTable} or \code{AddTable} respectively.}
}
\value{
A list containing the Dual STATIS results of the processing. This includes the bagplots
of Interstructure and Intrastructure for monitoring.
}
\description{
Default routine for automatic processing of reference batches and/or
testing batches using Dual STATIS for the DS-PC approach
}
\examples{
#DS-PC PlasticBags data
data( PlasticBags )
REFtab = PlasticBags$Ref$data
REFfac = PlasticBags$Ref$factor
process.results = AutoProcessing(RefTable = REFtab,FactOnTable = FALSE,RefFactor = REFfac )
process.results$REF$PC_proj$G
process.results$BPS$IS_bp
}
