% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{DualSTATIS}
\alias{DualSTATIS}
\title{Dual STATIS analysis for reference data}
\usage{
DualSTATIS(Tobj)
}
\arguments{
\item{Tobj}{The table object containing the \code{Xk_data} to  be preprocessed}
}
\value{
A list containing the Dual STATIS results, coded as \code{DSr}.
The projected coordinates are contained on the \code{PC_proj} list. The \code{G} matrix
has the projection of tables onto the Interstructure, the \code{V} matrix allows
new tables projection onto the Interstructure, the \code{F_scores} matrix
contains the principal components projections onto the Intrastructure, the \code{Fk} list
is composed by the partial factor scores matrix for every table, the \code{Pk} list
considers the projection matrices for every reference table and the \code{COj} list
contains the same partial factor scores as \code{Fk} organized by variable. The
\code{DSr} list, also has the \code{Xgsvd} list with the information of the generalized
singular value decomposition results. Additionally, original data from \code{Tobj} is
conserved
}
\description{
Performs Dual STATIS on a Table object derived from the \code{TableObject}
function. It provides all elements needed for analysis and monitoring
}
\examples{
#Iris Data
data(iris)
IrisTobj = TableObject(JointTable = iris[,1:4], TableFactor = iris[,5])
IrisDSr = DualSTATIS( IrisTobj )
IrisDSr$PC_proj$G
IrisDSr$PC_proj$COj

#DS-PC PlasticBags data
data( PlasticBags )
PBsTobj = TableObject( PlasticBags$Ref$data,  PlasticBags$Ref$factor)
PBsDSr = DualSTATIS( PreprocessTobj(PBsTobj)  )
PBsDSr$PC_proj$G
PBsDSr$PC_proj$COj
}
