% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{DualSTATISprojection}
\alias{DualSTATISprojection}
\title{Dual STATIS projection of new tables}
\usage{
DualSTATISprojection(Dsrs, newTobj)
}
\arguments{
\item{Dsrs}{The Dual STATIS results object containing the processsed reference information}

\item{newTobj}{The table object containing the prepared information of new tables}
}
\value{
A list containing the projection of new tables, coded as \code{nDSr}.
The projected coordinates are contained on the \code{PC_proj} list.
Additionally, original data from \code{newTobj} is conserved
}
\description{
Projects new tables from a Table object derived from the \code{TableObject}
function. It provides all elements needed for monitoring of new tables.
}
\examples{
#DS-PC PlasticBags data
data( PlasticBags )
PBsTobj = TableObject( PlasticBags$Ref$ data,  PlasticBags$Ref$factor )
Xbar =  colMeans(PBsTobj$Original_data); Desv = apply(PBsTobj$Original_data,2,sd)
PBsDSr = DualSTATIS( PreprocessTobj(PBsTobj)  )
newTobj = TableObject( PlasticBags$Add$data,  PlasticBags$Add$factor )
procTobj = PreprocessTobj(newTobj, centering = Xbar, scaling = Desv)
NewTabProj = DUAL_STATIS_projection( PBsDSr, procTobj )
NewTabProj$PC_proj$G
}
