% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{GenBagplots}
\alias{GenBagplots}
\title{Computing of bagplots for projected tables}
\usage{
GenBagplots(Dsrs)
}
\arguments{
\item{Dsrs}{A list containing Dual STATIS results obtained from \code{DualSTATIS} function,}
}
\value{
A list containing coordinates of points conforming the hull and center of
the bagplots ordered counterclockwise.
}
\description{
Computing of hull and center of bagplots for Interstructure and
variables in the Intrastructure contained in a Dual STATIS results list derived
from \code{DualSTATIS} function.
}
\examples{
#DS-PC PlasticBags data
data( PlasticBags )
PBsTobj = TableObject( PlasticBags$Ref$data,  PlasticBags$Ref$factor)
PBsDSr = DualSTATIS( PreprocessTobj(PBsTobj)  )
GenBagplots( PBsDSr )
}
