% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlasticBags.R
\docType{data}
\name{PlasticBags}
\alias{PlasticBags}
\title{Simulated dataset of industrial plastic bags fabrication}
\format{
A list containing data from 200 reference batches stacked in \code{PlasticBags$Ref$data}
and 8 additional testing batches stacked in \code{PlasticBags$Add$data}. Every table is
conformed by 50 observations of 3 variables. A factor is asociated to each data table.
}
\usage{
PlasticBags
}
\description{
Simulated dataset of industrial plastic bags fabrication
}
\keyword{datasets}
