% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{PreprocessTobj}
\alias{PreprocessTobj}
\title{Preprocessing of the Table Object}
\usage{
PreprocessTobj(Tobj, centering = TRUE, scaling = TRUE, normalizing = TRUE)
}
\arguments{
\item{Tobj}{The table object containing the \code{Xk_data} to  be preprocessed}

\item{centering}{either a logical indicating whether center the data or not (usig global mean), a
value to perform the centering or a vector indicating a value for every variable.}

\item{scaling}{either a logical indicating whether scale the data or not (using global standard deviations), a
value to perform the centering or a vector indicating a value for every variable.}

\item{normalizing}{When TRUE, data is normalized dividing by the corresponding number
of rows in each table. If FALSE, the tables are not normalized. Optionally, a function which
takes each table as argument can be given, the value obtained will be divided from each
corresponding table.}
}
\value{
The Table object (\code{Tobj}) where \code{Xk_data} is preprocessed.
}
\description{
Performs an optional (but preferred) preprocessing over the \code{Xk_data}
in a Table Object.
}
\examples{
data(iris)
IrisTobj = TableObject(JointTable = iris[,1:4], TableFactor = iris[,5])
}
