% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{SeqFactor}
\alias{SeqFactor}
\title{Create sequences over a factor}
\usage{
SeqFactor(FactorToName)
}
\arguments{
\item{FactorToName}{The factor from which names are created.}
}
\value{
A character vector of names, matching the length of \code{FactorToName}.
}
\description{
Use a factor to create a vector of
unique names for tables observations.
}
\examples{
fac = TabFactor(  c(5,7,3,4) )
SeqFactor( fac )

SeqFactor( c("A","A","A","B","B","B","C","C","C","C") )
}
