% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{TabFactor}
\alias{TabFactor}
\title{Create a Factor from a vector}
\usage{
TabFactor(n_values, mode = "obs", name = "T")
}
\arguments{
\item{n_values}{When \code{mode = "dim"}, \code{n_values} is a vector containing
the number of observations per table in \code{n_values[1]} and the number
of tables in \code{n_values[2]}. When \code{mode = "obs"}, n_values is a vector containing
the number of observations in each table.}

\item{mode}{It can be set up to "dimensions mode" when \code{mode = "dim"} or
"observations mode" when \code{mode = "obs"}(default).}

\item{name}{A character string used as prefix in the creation of the Factor.
It can be useful to distinguish different set of tables.}
}
\value{
A factor that can be used to splitting a matrix of values into a list of data tables.
}
\description{
Use number of tables and observations, or a vector
containing the number of observations in every table to create a
Factor used un tables splitting for DS-PC analysis.
}
\examples{
TabFactor( c(5,7) )
TabFactor( n_values = c(5,7), mode = "obs" )
TabFactor( n_values = c(5,7), mode = "dim" )
TabFactor( n_values = c(5,7,3,4) )
TabFactor( n_values = c(5,7,3,4), name = "W" )
}
