% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{TableObject}
\alias{TableObject}
\title{Create a Tables Object}
\usage{
TableObject(JointTable, TableFactor, ObsNames = NULL, VarNames = NULL)
}
\arguments{
\item{JointTable}{A matrix or data frame containing the K reference tables
stacked downwards one after another. Is preferred that its columns and rows have names.}

\item{TableFactor}{A factor with the same number of entries as rows are in
\code{JointTable}.It is used to define the different tables involved in the analysis.}

\item{ObsNames}{A vector containing names for each row in \code{JointTable} (optional).}

\item{VarNames}{A vector containing names for each column in \code{JointTable} (optional).}
}
\value{
A Table Object (a list) containing the separated tables
in \code{Xk_data}, the given table with names in \code{Original_data},
the \code{TableFactor} used to split the given table, the names
assigned to rows and columns in \code{VarNames} and \code{ObsNames},
the order of tables extracted from the factor, in \code{TablesOrder},
and the number of variables and observations in \code{Nvars} and \code{Nobs}
for computation purposes.
}
\description{
Uses a matrix or data frame to organize the main parameters
needed to perform DS-PC analysis.
}
\examples{
data(iris)
TableObject(JointTable = iris[,1:4], TableFactor = iris[,5])
TableObject(iris[,1:4], iris[,5], ObsNames = SeqFactor(iris[,5]) )
TableObject(iris[,1:4], iris[,5], SeqFactor(iris[,5]), VarNames = paste("V",1:4,sep="") )
}
